refactor(api): move CombinedCourseDetails model to its own class
