+++ api/src/quest/domain/models/CombinedCourse.js
@@ -1,17 +1,6 @@
 import Joi from 'joi';
 
-import {
-  CombinedCourseParticipationStatuses,
-  CombinedCourseStatuses,
-} from '../../../prescription/shared/domain/constants.js';
-import { config } from '../../../shared/config.js';
 import { EntityValidationError } from '../../../shared/domain/errors.js';
-import { cryptoService as injectedCryptoService } from '../../../shared/domain/services/crypto-service.js';
-import { COMBINED_COURSE_ITEM_TYPES, CombinedCourseItem } from './CombinedCourseItem.js';
-import { CombinedCourseParticipationDetails } from './CombinedCourseParticipationDetails.js';
-import { DataForQuest } from './DataForQuest.js';
-import { Eligibility } from './Eligibility.js';
-import { TYPES } from './Requirement.js';
 
 const schema = Joi.object({
   id: Joi.number().allow(null),
@@ -74,271 +63,5 @@ export class CombinedCourse {
     if (error) {
       throw EntityValidationError.fromJoiErrors(error.details, undefined, { data: combinedCourse });
     }
-  }
-}
-
-export class CombinedCourseDetails extends CombinedCourse {
-  items = null;
-  #combinedCourseUrl = null;
-
-  constructor(
-    { id, code, organizationId, name, description, illustration, questId },
-    quest,
-    cryptoService = injectedCryptoService,
-  ) {
-    super({ id, code, organizationId, name, description, illustration, questId }, quest);
-    this.campaigns = [];
-    this.recommendableModuleIds = [];
-    this.cryptoService = cryptoService;
-  }
-
-  async setEncryptedUrl() {
-    this.#combinedCourseUrl = await this.cryptoService.encrypt('/parcours/' + this.code, config.module.secret);
-  }
-
-  setRecommandableModuleIds(recommendableModuleIds) {
-    this.recommendableModuleIds = recommendableModuleIds;
-  }
-
-  setItems({ campaigns, modules }) {
-    this.campaigns = campaigns;
-    this.modules = modules;
-  }
-
-  get hasCampaigns() {
-    return this.campaignIds.length > 0;
-  }
-
-  get hasModules() {
-    return this.moduleIds.length > 0;
-  }
-
-  get campaignIds() {
-    return this.quest.successRequirements
-      .filter((requirement) => requirement.requirement_type === TYPES.OBJECT.CAMPAIGN_PARTICIPATIONS)
-      .map(({ data }) => data.campaignId.data);
-  }
-
-  get moduleIds() {
-    return this.quest.successRequirements
-      .filter((requirement) => requirement.requirement_type === TYPES.OBJECT.PASSAGES)
-      .map(({ data }) => data.moduleId.data);
-  }
-
-  get participationDetails() {
-    return new CombinedCourseParticipationDetails({
-      id: this.participation.id,
-      status: this.status,
-      firstName: this.participation.firstName,
-      lastName: this.participation.lastName,
-      division: this.participation.division,
-      group: this.participation.group,
-      createdAt: this.participation.createdAt,
-      updatedAt: this.participation.updatedAt,
-      hasFormationItem: this.items.some(({ type }) => type === COMBINED_COURSE_ITEM_TYPES.FORMATION),
-      nbCampaigns: this.items.filter(({ type }) => type === COMBINED_COURSE_ITEM_TYPES.CAMPAIGN).length,
-      nbModules: this.items.filter(
-        ({ type }) => type === COMBINED_COURSE_ITEM_TYPES.MODULE || type === COMBINED_COURSE_ITEM_TYPES.FORMATION,
-      ).length,
-      nbModulesCompleted: this.items.filter(
-        ({ isCompleted, type }) => isCompleted && type === COMBINED_COURSE_ITEM_TYPES.MODULE,
-      ).length,
-      nbCampaignsCompleted: this.items.filter(
-        ({ isCompleted, type }) => isCompleted && type === COMBINED_COURSE_ITEM_TYPES.CAMPAIGN,
-      ).length,
-    });
-  }
-
-  #createCampaignCombinedCourseItem({
-    campaign,
-    participationStatus,
-    isCompleted,
-    isLocked,
-    masteryRate,
-    totalStagesCount,
-    validatedStagesCount,
-  }) {
-    return new CombinedCourseItem({
-      id: campaign.id,
-      reference: campaign.code,
-      title: campaign.title,
-      type: COMBINED_COURSE_ITEM_TYPES.CAMPAIGN,
-      masteryRate: isCompleted ? masteryRate : null,
-      participationStatus,
-      isCompleted,
-      isLocked,
-      totalStagesCount: isCompleted ? totalStagesCount : null,
-      validatedStagesCount: isCompleted ? validatedStagesCount : null,
-    });
-  }
-
-  #createModuleCombinedCourseItem(module, participationStatus, isCompleted, isLocked) {
-    return new CombinedCourseItem({
-      id: module.id,
-      reference: module.slug,
-      title: module.title,
-      type: COMBINED_COURSE_ITEM_TYPES.MODULE,
-      redirection: this.#combinedCourseUrl,
-      participationStatus,
-      isCompleted,
-      isLocked,
-      duration: module?.duration,
-      image: module.image,
-    });
-  }
-
-  #createFormationCombinedCourseItem(targetProfileId) {
-    return new CombinedCourseItem({
-      id: 'formation_' + this.quest.id + '_' + targetProfileId,
-      reference: targetProfileId,
-      type: COMBINED_COURSE_ITEM_TYPES.FORMATION,
-    });
-  }
-
-  #createFormationCombinedCourseItemIfNeeded(recommendableModule, targetProfileIdsThatNeedAFormationItem) {
-    const targetProfileId = targetProfileIdsThatNeedAFormationItem.find((targetProfileIdOfCampaign) =>
-      recommendableModule.targetProfileIds.includes(targetProfileIdOfCampaign),
-    );
-
-    const shouldBeInFormationItem = Boolean(targetProfileId);
-    const hasFormationItem = this.items.find((item) => {
-      if (item.type !== COMBINED_COURSE_ITEM_TYPES.FORMATION) return false;
-      if (item.reference === targetProfileId) return true;
-      return false;
-    });
-    if (shouldBeInFormationItem) {
-      if (!hasFormationItem) {
-        return this.#createFormationCombinedCourseItem(targetProfileId);
-      } else {
-        return undefined;
-      }
-    }
-  }
-
-  #isCombinedCourseItemLocked(previousItem) {
-    if (!previousItem) {
-      return false;
-    }
-    if (previousItem.isLocked) {
-      return true;
-    } else {
-      return previousItem.isCompleted ? false : true;
-    }
-  }
-
-  updateItemsFromPassages(passages) {
-    const updatedDataForQuest = new DataForQuest({
-      eligibility: new Eligibility({
-        organizationLearner: this.dataForQuest.eligibility.organizationLearner,
-        organization: this.dataForQuest.eligibility.organization,
-        campaignParticipations: this.dataForQuest.eligibility.campaignParticipations,
-        passages,
-      }),
-    });
-    this.#generateItems({
-      dataForQuest: updatedDataForQuest,
-      participation: this.participation,
-    });
-    return this;
-  }
-
-  #setParticipationStatus(participation) {
-    this.participation = participation;
-
-    if (!this.participation) {
-      this.status = CombinedCourseStatuses.NOT_STARTED;
-    } else {
-      this.status =
-        participation.status === CombinedCourseParticipationStatuses.STARTED
-          ? CombinedCourseStatuses.STARTED
-          : CombinedCourseStatuses.COMPLETED;
-    }
-  }
-
-  #generateItems({ dataForQuest } = {}) {
-    this.items = [];
-    this.dataForQuest = dataForQuest;
-
-    const targetProfileIdsThatNeedAFormationItem = [];
-
-    for (const requirement of this.quest.successRequirements) {
-      const previousItem = this.items[this.items.length - 1];
-      const isLocked = this.#isCombinedCourseItemLocked(previousItem);
-
-      if (requirement.requirement_type === TYPES.OBJECT.CAMPAIGN_PARTICIPATIONS) {
-        const isCompleted = dataForQuest ? requirement.isFulfilled(dataForQuest) : false;
-        const campaign = this.campaigns.find((item) => item.id === requirement.data.campaignId.data);
-        const associatedParticipation = dataForQuest?.campaignParticipations?.find(
-          (campaignParticipation) => campaignParticipation.campaignId === campaign.id,
-        );
-
-        const participationStatus = associatedParticipation?.status;
-
-        const doesCampaignRecommendModules =
-          this.recommendableModuleIds.find((recommendableModule) => {
-            if (!this.moduleIds.includes(recommendableModule.moduleId)) return false;
-            return recommendableModule.targetProfileIds.includes(campaign.targetProfileId);
-          }) ?? [];
-
-        if (doesCampaignRecommendModules && !isCompleted) {
-          targetProfileIdsThatNeedAFormationItem.push(campaign.targetProfileId);
-        }
-
-        this.items.push(
-          this.#createCampaignCombinedCourseItem({
-            campaign,
-            participationStatus,
-            isCompleted,
-            isLocked,
-            masteryRate: associatedParticipation?.masteryRate,
-            validatedStagesCount: associatedParticipation?.validatedStagesCount,
-            totalStagesCount: associatedParticipation?.totalStagesCount,
-          }),
-        );
-      } else if (requirement.requirement_type === TYPES.OBJECT.PASSAGES) {
-        const isCompleted = dataForQuest ? requirement.isFulfilled(dataForQuest) : false;
-        const module = this.modules.find((item) => item.id === requirement.data.moduleId.data);
-        const passage = dataForQuest?.passages.find((passage) => passage.moduleId === module.id);
-        const participationStatus = passage?.status;
-
-        const recommendableModule = this.recommendableModuleIds.find(
-          (potentiallyRecommendedModule) => potentiallyRecommendedModule.moduleId === module.id,
-        );
-        const isRecommandable = Boolean(recommendableModule);
-
-        if (!isRecommandable) {
-          this.items.push(this.#createModuleCombinedCourseItem(module, participationStatus, isCompleted, isLocked));
-          continue;
-        }
-
-        const isRecommended = this.recommendedModuleIdsForUser.find(
-          (recommendedModule) => recommendedModule.moduleId === module.id,
-        );
-
-        if (isRecommended) {
-          this.items.push(this.#createModuleCombinedCourseItem(module, participationStatus, isCompleted, isLocked));
-          continue;
-        }
-
-        const formationCombinedCourseItem = this.#createFormationCombinedCourseItemIfNeeded(
-          recommendableModule,
-          targetProfileIdsThatNeedAFormationItem,
-        );
-
-        if (formationCombinedCourseItem) {
-          this.items.push(formationCombinedCourseItem);
-        }
-      }
-    }
-  }
-
-  setDataAndGenerateItems({ recommendedModuleIdsForUser = [], dataForQuest, participation = null } = {}) {
-    this.recommendedModuleIdsForUser = recommendedModuleIdsForUser;
-    this.#setParticipationStatus(participation);
-    this.#generateItems({ dataForQuest });
-  }
-
-  isSuccessful() {
-    return this.quest.isSuccessful(this.dataForQuest);
   }
 }
