+++ api/tests/quest/integration/domain/usecases/get-combined-course-by-id_test.js
@@ -1,10 +1,10 @@
-import { CombinedCourseDetails } from '../../../../../src/quest/domain/models/CombinedCourse.js';
 import { CombinedCourseBlueprint } from '../../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
 import { usecases } from '../../../../../src/quest/domain/usecases/index.js';
 import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
 import { catchErr } from '../../../../tooling/test-utils/error.js';
+import { CombinedCourseDetails } from '../../../../../src/quest/domain/models/CombinedCourseDetails.js';
 
 describe('Quest | Integration | Domain | Usecases | getCombinedCourseById', function () {
   it('should return a CombinedCourseDetails instance with quest and combined course data', async function () {
