+++ api/tests/tooling/domain-builder/factory/build-combined-course-details.js
@@ -1,10 +1,11 @@
 import { Campaign } from '../../../../src/quest/domain/models/Campaign.js';
-import { CombinedCourse, CombinedCourseDetails } from '../../../../src/quest/domain/models/CombinedCourse.js';
+import { CombinedCourse } from '../../../../src/quest/domain/models/CombinedCourse.js';
 import { CombinedCourseBlueprint } from '../../../../src/quest/domain/models/CombinedCourseBlueprint.js';
 import { DataForQuest } from '../../../../src/quest/domain/models/DataForQuest.js';
 import { Eligibility } from '../../../../src/quest/domain/models/Eligibility.js';
 import { Module } from '../../../../src/quest/domain/models/Module.js';
 import { Quest } from '../../../../src/quest/domain/models/Quest.js';
+import { CombinedCourseDetails } from '../../../../src/quest/domain/models/CombinedCourseDetails.js';
 
 function buildCombinedCourse({ name, code, organizationId, questId } = {}) {
   return new CombinedCourse({
